+++ modules/@apostrophecms/task/index.js
@@ -117,6 +117,8 @@ module.exports = {
       // normally be hyphenated, i.e. at the command line you might write
       // `--total=20`.
       //
+      // This method will return the value returned by the task.
+      //
       // **Gotchas**
       //
       // If you can invoke a method directly rather than invoking a task, do
@@ -132,13 +134,14 @@ module.exports = {
       // task developer might assume they can exit the process directly.
 
       async invoke(name, args, options) {
+        let result;
         const telemetry = self.apos.telemetry;
         const spanName = `task:${self.__meta.name}:${name}`;
+        const aposArgv = self.apos.argv;
         await telemetry.startActiveSpan(spanName, async (span) => {
           span.setAttribute(SemanticAttributes.CODE_FUNCTION, 'invoke');
           span.setAttribute(SemanticAttributes.CODE_NAMESPACE, '@apostrophecms/task');
           try {
-            const aposArgv = self.apos.argv;
             if (Array.isArray(args)) {
               args.splice(0, 0, name);
             } else {
@@ -155,16 +158,17 @@ module.exports = {
             };
             span.setAttribute(telemetry.Attributes.ARGV, telemetry.stringify(argv));
             self.apos.argv = argv;
-            await task.task(argv);
-            self.apos.argv = aposArgv;
+            result = await task.task(argv);
             span.setStatus({ code: telemetry.api.SpanStatusCode.OK });
           } catch (err) {
             telemetry.handleError(span, err);
             throw err;
           } finally {
+            self.apos.argv = aposArgv;
             span.end();
           }
         });
+        return result;
       },
 
       // Identifies the task corresponding to the given command line argument.